refactor: svg 폴더 이동
