+++ src/app/(afterlogin)/upload/[id]/_component/_svgs/InputFormSvgs.module.scss
