+++ src/app/(afterlogin)/upload/[id]/_component/_svgs/PptImgSvgs.tsx
